Create README.md
Dear Colleagues
My name is Alexander and I want to create a TTS for the Georgian language. I have about 30 hours of material (audio and corresponding text) presented in the form required for Takotrona 2. My goal is to create a TTS for the Georgian language. I would like to use a pre-trained neural network but I have some questions. Since the Georgian language has its own original writing and alphabet, I don’t know in which folder the Georgian alphabet should be placed. What else is required for the preparation and subsequent training of a neural network in order to teach it TTS. Since I am not a programmer, maybe someone can help me and provide step-by-step instructions (for a non-specialist) with which I could make TTS for the Georgian language.